refactor clone() into base class so it can be used in any model
